add PAWSS tracker

+++ src/Features/PatchGrayGFeature.cpp
@@ -1,7 +1,7 @@
-#include "Features/PatchGrayGFeature.h"
-#include "mUtils.h"
-#include "ImageRep.h"
-#include "Sample.h"
+#include <PAWSS/Features/PatchGrayGFeature.h>
+#include <PAWSS/mUtils.h>
+#include <PAWSS/ImageRep.h>
+#include <PAWSS/Sample.h>
 
 static const int kMiniPatchRadius = 3;
 
@@ -109,7 +109,7 @@ void PatchGrayGFeature::PrepEval(const multiSample &samples)
         cv::integral(hists[i+mGrayBinNum], mIntegs[i+mGrayBinNum], CV_32F);
 
 
-    for(int i=0; i<hists.size(); ++i)
+    for(size_t i=0; i<hists.size(); ++i)
         hists[i].release();
     hists.clear();
 }
