+++ src/Features/PatchHsvFeature.cpp
@@ -1,7 +1,7 @@
-#include "Features/PatchHsvFeature.h"
-#include "Sample.h"
-#include "mUtils.h"
-#include "ImageRep.h"
+#include <PAWSS/Features/PatchHsvFeature.h>
+#include <PAWSS/mUtils.h>
+#include <PAWSS/ImageRep.h>
+#include <PAWSS/Sample.h>
 
 static const int kMiniPatchRadius = 1;
 
@@ -100,7 +100,7 @@ void PatchHsvFeature::PrepEval(const multiSample &samples)
         cv::integral(hists[i], mIntegs[i], CV_32F);
 //    cv::integral(weight_map, mWeightInteg, CV_64F);
 
-    for(int i=0; i<hists.size(); ++i)
+    for(size_t i=0; i<hists.size(); ++i)
         hists[i].release();
     hists.clear();
 }
