+++ src/Features/PatchHsvGFeature.cpp
@@ -1,7 +1,7 @@
-#include "Features/PatchHsvGFeature.h"
-#include "Sample.h"
-#include "mUtils.h"
-#include "ImageRep.h"
+#include <PAWSS/Features/PatchHsvGFeature.h>
+#include <PAWSS/mUtils.h>
+#include <PAWSS/ImageRep.h>
+#include <PAWSS/Sample.h>
 
 static const int kMiniPatchRadius = 3;
 
@@ -125,7 +125,7 @@ void PatchHsvGFeature::PrepEval(const multiSample &samples)
         cv::integral(hists[i+mHsvBinNum], mIntegs[i+mHsvBinNum], CV_32F);
 
 
-    for(int i=0; i<hists.size(); ++i)
+    for(size_t i=0; i<hists.size(); ++i)
         hists[i].release();
     hists.clear();
 }
