+++ src/Features/PatchHsvMFeature.cpp
@@ -1,7 +1,7 @@
-#include "Features/PatchHsvMFeature.h"
-#include "mUtils.h"
-#include "ImageRep.h"
-#include "Sample.h"
+#include <PAWSS/Features/PatchHsvMFeature.h>
+#include <PAWSS/mUtils.h>
+#include <PAWSS/ImageRep.h>
+#include <PAWSS/Sample.h>
 
 static const int kMiniPatchRadius = 1;
 
@@ -121,7 +121,7 @@ void PatchHsvMFeature::PrepEval(const multiSample &samples)
         cv::integral(hists[i+mHsvBinNum], mIntegs[i+mHsvBinNum], CV_32F);
 
 
-    for(int i=0; i<hists.size(); ++i)
+    for(size_t i=0; i<hists.size(); ++i)
         hists[i].release();
     hists.clear();
